requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
numpy>=1.26.0
Pillow>=10.0.0
setuptools>=68.0.0
wheel>=0.40.0
